requirements.txt
fastapi
uvicorn[standard]
httpx[http2]
motor
python-dotenv
pyjwt[crypto]
passlib[bcrypt]
python-multipart